+++ R/helpers.R
@@ -1,0 +1,5 @@
+cancensus_base_url <- function(){
+  url <- getOption("cancensus.base_url")
+  if (is.null(url)) url <- "https://censusmapper.ca"
+  url
+}